Move method getCurrentPermissions into RolesTests class and adjust it to work for LightAdminRoles... as well as LightAdminPrivileges...
